+++ enterprise/litellm_enterprise/proxy/vector_stores/endpoints.py
@@ -31,129 +31,6 @@ from litellm.vector_stores.vector_store_registry import VectorStoreRegistry
 router = APIRouter()
 
 ########################################################
-# OpenAI Compatible Endpoints
-########################################################
-
-@router.post("/v1/vector_stores/{vector_store_id}/search", dependencies=[Depends(user_api_key_auth)])
-@router.post("/vector_stores/{vector_store_id}/search", dependencies=[Depends(user_api_key_auth)])
-async def vector_store_search(
-    request: Request,
-    vector_store_id: str,
-    fastapi_response: Response,
-    user_api_key_dict: UserAPIKeyAuth = Depends(user_api_key_auth),
-):
-    """
-    Search a vector store.
-
-    API Reference:
-    https://platform.openai.com/docs/api-reference/vector-stores/search
-    """
-    from litellm.proxy.proxy_server import (
-        _read_request_body,
-        general_settings,
-        llm_router,
-        proxy_config,
-        proxy_logging_obj,
-        select_data_generator,
-        user_api_base,
-        user_max_tokens,
-        user_model,
-        user_request_timeout,
-        user_temperature,
-        version,
-    )
-
-    data = await _read_request_body(request=request)
-    if "vector_store_id" not in data:
-        data["vector_store_id"] = vector_store_id
-    processor = ProxyBaseLLMRequestProcessing(data=data)
-    try:
-        return await processor.base_process_llm_request(
-            request=request,
-            fastapi_response=fastapi_response,
-            user_api_key_dict=user_api_key_dict,
-            route_type="vector_store_search",
-            proxy_logging_obj=proxy_logging_obj,
-            llm_router=llm_router,
-            general_settings=general_settings,
-            proxy_config=proxy_config,
-            select_data_generator=select_data_generator,
-            model=None,
-            user_model=user_model,
-            user_temperature=user_temperature,
-            user_request_timeout=user_request_timeout,
-            user_max_tokens=user_max_tokens,
-            user_api_base=user_api_base,
-            version=version,
-        )
-    except Exception as e:
-        raise await processor._handle_llm_api_exception(
-            e=e,
-            user_api_key_dict=user_api_key_dict,
-            proxy_logging_obj=proxy_logging_obj,
-            version=version,
-        )
-
-
-
-@router.post("/v1/vector_stores", dependencies=[Depends(user_api_key_auth)])
-@router.post("/vector_stores", dependencies=[Depends(user_api_key_auth)])
-async def vector_store_create(
-    request: Request,
-    fastapi_response: Response,
-    user_api_key_dict: UserAPIKeyAuth = Depends(user_api_key_auth),
-):
-    """
-    Create a vector store.
-
-    API Reference:
-    https://platform.openai.com/docs/api-reference/vector-stores/create
-    """
-    from litellm.proxy.proxy_server import (
-        _read_request_body,
-        general_settings,
-        llm_router,
-        proxy_config,
-        proxy_logging_obj,
-        select_data_generator,
-        user_api_base,
-        user_max_tokens,
-        user_model,
-        user_request_timeout,
-        user_temperature,
-        version,
-    )
-
-    data = await _read_request_body(request=request)
-    processor = ProxyBaseLLMRequestProcessing(data=data)
-    try:
-        return await processor.base_process_llm_request(
-            request=request,
-            fastapi_response=fastapi_response,
-            user_api_key_dict=user_api_key_dict,
-            route_type="vector_store_create",
-            proxy_logging_obj=proxy_logging_obj,
-            llm_router=llm_router,
-            general_settings=general_settings,
-            proxy_config=proxy_config,
-            select_data_generator=select_data_generator,
-            model=None,
-            user_model=user_model,
-            user_temperature=user_temperature,
-            user_request_timeout=user_request_timeout,
-            user_max_tokens=user_max_tokens,
-            user_api_base=user_api_base,
-            version=version,
-        )
-    except Exception as e:
-        raise await processor._handle_llm_api_exception(
-            e=e,
-            user_api_key_dict=user_api_key_dict,
-            proxy_logging_obj=proxy_logging_obj,
-            version=version,
-        )
-
-########################################################
 # Management Endpoints
 ########################################################
 @router.post(
